Rats cumulative angle
Created a function that determines the rats cumulative angle from its starting point. A corrsponding graphing method was also created.

Futhermore, vl_read was updated to take the position of the target that is being chased and also the clockwiseness of the rat. The label occassionally disagrees with my cross product calculation, so something needs to be checked. Perhaps the vl label uses the velocity vector, while my calculation only uses position?

Lastly, the position of the target is visualized in the EnvironmentReader portion of LocationAnimation.py